requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.7
# by the following command:
#
#    pip-compile
#
appdirs==1.4.4
    # via wandb
certifi==2022.12.7
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via wandb
cycler==0.11.0
    # via matplotlib
docker-pycreds==0.4.0
    # via wandb
fonttools==4.38.0
    # via matplotlib
gitdb==4.0.10
    # via gitpython
gitpython==3.1.31
    # via wandb
idna==3.4
    # via requests
imageio==2.28.1
    # via scikit-image
importlib-metadata==6.6.0
    # via click
joblib==1.2.0
    # via scikit-learn
kiwisolver==1.4.4
    # via matplotlib
matplotlib==3.5.3
    # via -r requirements.in
networkx==2.6.3
    # via scikit-image
nibabel==4.0.2
    # via -r requirements.in
numpy==1.21.6
    # via
    #   -r requirements.in
    #   imageio
    #   matplotlib
    #   nibabel
    #   opencv-python
    #   pandas
    #   pylibjpeg
    #   pylibjpeg-libjpeg
    #   pywavelets
    #   scikit-image
    #   scikit-learn
    #   scipy
    #   tifffile
    #   torchvision
nvidia-cublas-cu11==11.10.3.66
    # via
    #   nvidia-cudnn-cu11
    #   torch
nvidia-cuda-nvrtc-cu11==11.7.99
    # via torch
nvidia-cuda-runtime-cu11==11.7.99
    # via torch
nvidia-cudnn-cu11==8.5.0.96
    # via torch
opencv-python==4.7.0.72
    # via -r requirements.in
packaging==23.1
    # via
    #   matplotlib
    #   nibabel
    #   scikit-image
pandas==1.3.5
    # via -r requirements.in
pathtools==0.1.2
    # via wandb
pillow==9.5.0
    # via
    #   -r requirements.in
    #   imageio
    #   matplotlib
    #   scikit-image
    #   torchvision
protobuf==4.22.4
    # via wandb
psutil==5.9.5
    # via wandb
pydicom==2.3.1
    # via -r requirements.in
pylibjpeg==1.4.0
    # via -r requirements.in
pylibjpeg-libjpeg==1.3.4
    # via -r requirements.in
pyparsing==3.0.9
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
python-gdcm==3.0.21
    # via -r requirements.in
pytz==2023.3
    # via pandas
pywavelets==1.3.0
    # via scikit-image
pyyaml==6.0
    # via wandb
requests==2.30.0
    # via
    #   torchvision
    #   wandb
scikit-image==0.19.3
    # via -r requirements.in
scikit-learn==1.0.2
    # via -r requirements.in
scipy==1.7.3
    # via
    #   scikit-image
    #   scikit-learn
sentry-sdk==1.22.1
    # via wandb
setproctitle==1.3.2
    # via wandb
six==1.16.0
    # via
    #   docker-pycreds
    #   python-dateutil
smmap==5.0.0
    # via gitdb
threadpoolctl==3.1.0
    # via scikit-learn
tifffile==2021.11.2
    # via scikit-image
torch==1.13.1
    # via
    #   -r requirements.in
    #   torchvision
torchvision==0.14.1
    # via -r requirements.in
typing-extensions==4.5.0
    # via
    #   gitpython
    #   importlib-metadata
    #   kiwisolver
    #   torch
    #   torchvision
    #   wandb
urllib3==1.26.15
    # via
    #   requests
    #   sentry-sdk
wandb==0.15.2
    # via -r requirements.in
wheel==0.40.0
    # via
    #   nvidia-cublas-cu11
    #   nvidia-cuda-runtime-cu11
zipp==3.15.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
